requirements.txt
openai==1.55.3
httpx== 0.27.2
streamlit==1.40.1
python-docx==0.8.11
python-dotenv==1.0.0
PyPDF2==3.0.1